+++ packages/dataviz/src/components/BarChart/VerticalBarChart/VerticalBarChart.component.test.tsx
@@ -1,4 +1,5 @@
 import { render } from '@testing-library/react';
+
 import VerticalBarChart from './VerticalBarChart.component';
 
 describe('Vertical bar chart', () => {