+++ tools/babel-plugin-assets-api/__fixtures__/first/output.js
@@ -5,9 +5,9 @@ import React from 'react';
 
 import assetsAPI from '@talend/assets-api';
 
-assetsAPI.getURL('/', 'react', '18.2.0');
+assetsAPI.getURL('/', 'react', '18.3.1');
 assetsAPI.getURL('/', 'react', '16.13.0');
-assetsAPI.getJSON('/foo.json', 'react', '18.2.0');
+assetsAPI.getJSON('/foo.json', 'react', '18.3.1');
 assetsAPI.getJSON('/foo.json', 'react', '16.13.0');
-assetsAPI.getUMD('react', '18.2.0', 'React', '/umd/react.production.min.js');
+assetsAPI.getUMD('react', '18.3.1', 'React', '/umd/react.production.min.js');
 assetsAPI.getUMD('react', '16.13.0', 'ReactWrong', '/umd/react.production.min.js');