+++ tools/babel-plugin-import-from-index/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @talend/babel-plugin-import-from-index
 
+## 1.8.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
 ## 1.7.1
 
 ### Patch Changes