+++ tools/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @talend/eslint-plugin
 
+## 1.3.1
+
+### Patch Changes
+
+- f1c2fdf: Add `btn-inverse` to list of warning for eslint bootstrap classes
+
+## 1.3.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/scripts-config-cdn@10.12.0
+
 ## 1.2.0
 
 ### Minor Changes