+++ tools/scripts-config-storybook-lib/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @talend/scripts-config-storybook-lib
 
+## 5.4.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/scripts-config-react-webpack@16.8.0
+  - @talend/dynamic-cdn-webpack-plugin@14.0.0
+
+## 5.3.2
+
+### Patch Changes
+
+- 3bd16fc: Make it compatible with ESM
+- Updated dependencies [3bd16fc]
+  - @talend/scripts-config-react-webpack@16.7.0
+
 ## 5.3.1
 
 ### Patch Changes